requirements.txt
supabase-py
requests
beautifulsoup4
feedparser
python-dateutil
apify-client
